fix(services,handlers): opcode 93 inerte, soin borne, position de grille
SkyUpgradeItemHandler devient un no-op journalise : la ligne REGWORK1 de
l'opcode 93 (P_SKY_UP_ITEM_SEND) est enfermee dans un #ifdef __REBIRTH__ mort
(S04_MyWork01.cpp:97-99), donc le serveur legacy coupe la session a reception.
Lui donner un vrai comportement ouvrirait une surface que le client n'utilise
pas.

UseInventoryItemService levait une ArgumentException atteignable quand le soin
calcule depassait le manque de vie : borne au site plutot que par un ecretage
global, qui aurait masque le probleme ailleurs.

HotkeyActionService : le retrait appliquait la position de grille (0,0) au lieu
de celle demandee par le client, l'item revenait donc toujours dans le coin.

+++ src/Fenrir.Application.Game.Services/ItemModification/CraftItemService.cs
@@ -154,8 +154,8 @@ public sealed class CraftItemService(
         var freeSlot = InventoryFreeSlotFinder.Find(state.Inventory, worldData,
             CraftRecipeCatalog.AdvancedElixirResultBaseItemId, state.InventoryDate, GameDate.Today());
         var hasFreeSlot = freeSlot is not null;
-        var resultPage = freeSlot?.Container ?? (byte)0;
-        var resultIndex = freeSlot?.Slot ?? (byte)0;
+        var resultPage = freeSlot?.Container ?? 0;
+        var resultIndex = freeSlot?.Slot ?? 0;
 
         var resolved = CraftResolver.ResolveAdvancedElixir(material, hasFreeSlot, SystemRandomSource.Instance);
